requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.3
soupsieve==2.5
google-generativeai==0.4.1
reportlab==4.1.0
python-multipart==0.0.9
tenacity==8.2.3
urllib3==2.2.1
brotli==1.1.0
pdfminer.six==20231228
docx2txt==0.8
chardet==5.2.0
Jinja2==3.1.3
aiofiles==23.2.1
